add style to select boxes that have the "size" property
Style for <select multiple="multiple"> was already similar to regular input
boxes, but it has an implicit size=5 attribute. When select boxes are created
with explicitely set size=X all the nice styles of select[multiple] are not
being used.

in Chrome to make select boxes apply padding (like input boxes) it was also
required to @include appearance(none); as per stackoverflow.com/a/27508479/11414