Simplify the test
The test currently runs `bitters install`, which outputs Bitters' Sass files to
compile against, but it's doing so by using whatever local version of Bitters
you have installed, not what is represented in the current state of the repo. So
if you make a breaking change to the Sass, this test will not capture that.

This commit removes the `bitters install` command and instead points directly to
the Sass partials locally within the repo. The test will then try to compile
Bitters' Sass and if it doesn't compile, the test fails.

In the future, we should add more robust tests to specifically test the Bitters
CLI.