Move spacing variable to `space()` function.
Traditionally we have had spacing handled by a series of variables, i.e.
`$base-spacing`. This change adds the `space()` function. This function uses the
Bourbon `modular-scale()` function to create a series of values based on the
global modular scale multiplied by the `$base-line-height`, returning a `rem`
value. This change allows for a high level of spacing granularity while
maintaining a simple and scaleable approach, without polluting the global
variable namespace.