Remove disabling of link text decoration
Styling links only through color and removing their default underline
is an accessibility concern for people with color blindness. It's
possible they would not be able to discern the difference between
regular text and a link.

This commit also sets the `text-decoration-skip` property to `ink` for
all links, which only draws text decoration where it does not touch or
closely approach a glyph.