Disconnect $base-spacing from $base-line-height
The two serve different purposes and their values don't need to be
relational. Also, since we do not facilitate or support a vertical
rhythm, the two variables do not need to be connected.

One situation where I've found this to be problematic (when they are
connected) is when I want to tweak the line height for a specific
typeface, but I don't want that change to manipulate spacing throughout
the whole application.