Refactor background color variables
`$base-background-color` is only used for setting the background color
of form inputs. That variable name doesn't make sense in that context;
it makes it seem as though you can set this variable and change the
background color of the viewport.

This change swaps the name of `$base-background-color` to
`$viewport-background-color` and sets it as a value on the `html`
element. It also removes the `$secondary-background-color` variable,
which is not used at all.

Finally, a specific and private `$_form-background-color` variable is
added to set the background color on form inputs.