Replace gulp with Parcel
We have a small ‘contributing’ site that uses gulp as a build process.
It compiles Sass and runs a little server. It requires a bit of
configuration that we have to maintain.

This commit swaps gulp out for Parcel to reduce the code we need to
maintain, as Parcel is a zero-configuration bundler.

Closes https://github.com/thoughtbot/bitters/issues/351