Deploy contributing site to Netlify
This add configuration to deploy the contributing site to Netlify so
that we can get deploy previews for each GitHub Pull Request. This makes
it easier to review code and visual changes.

Since we don't intend for this contributing site to be public-facing,
a redirect is added to point the Netlify site to the Bitters GitHub
repo.

Netlify documentation on `netlify.toml`:
https://docs.netlify.com/configure-builds/file-based-configuration/

Closes: https://github.com/thoughtbot/bitters/issues/355