Unify stylelint version, ignore from Dependabot
This locks stylelint to specific versions and ignore it from Dependabot
so that it doesn't automatically open Pull Requests to update them.

The reason is because Hound has a specific version of stylelint that it
runs, and so while we might be using v12 of stylelint locally, Hound is
running v10 in our Pull Requests and so the results can differ.

Hound publishes their [supported linter versions][supported-linters]
and we can configure Hound to use specific versions via its `.hound.yml`
config.

Since these are just development tools, I think its probably fine for
them to be updated manually and less frequent than with Dependabot.

[supported-linters]: http://help.houndci.com/en/articles/2461415-supported-linters